pyproject.toml
[project]
name = "bread_fast_bot"
version = "0.1.0"
description = "Bot for bread ordering"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.5.2",
    "pydantic[dotenv]>=2.9.2",
    "python-telegram-bot>=21.5",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
]
